segment_requirements.txt
tkinter
pillow
numpy
glob
opencv-python